[NOP] Rename union -> sketch because union is a c++ keyword.
